requirements.txt
google-generativeai>=0.3.0
mcp>=0.8.0
aiohttp>=3.8.4
scholarly>=1.7.0
semantic-scholar-api>=1.0.0
pyvis>=0.3.1
networkx>=3.1
beautifulsoup4>=4.11.1
bibtexparser>=1.4.0
python-dotenv>=1.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
